Cast pwm to ints

+++ compute/io/serial_to_arduino.py
@@ -59,8 +59,8 @@ def zmq_to_serial(context: zmq.Context, uart_port: str):
             elif left < 0 and right < 0:
                 directions - 3
 
-            left_pwm = min(PWM_MAX, (abs(left) * K / MAX_VOLTAGE) * PWM_MAX)  # Get PWM, clip at PWM_MAX
-            right_pwm = min(PWM_MAX, (abs(right) * K / MAX_VOLTAGE) * PWM_MAX)  # Get PWM, clip at PWM_MAX
+            left_pwm = int(min(PWM_MAX, (abs(left) * K / MAX_VOLTAGE) * PWM_MAX))  # Get PWM, clip at PWM_MAX
+            right_pwm = int(min(PWM_MAX, (abs(right) * K / MAX_VOLTAGE) * PWM_MAX))  # Get PWM, clip at PWM_MAX
             match directions:
                 case 0:
                     print(f"Pushing - Left: {left_pwm}, Right: {right_pwm}")
